fix: move download and enable button to outside of details tab
